+++ CHANGELOG.md
@@ -6,12 +6,15 @@ All notable changes to this project will be documented in this file.
 
 ### Upgrade Notes
 
+- SCC v1 Terraform has been deprecated. SCC v1 resources can still be managed from `/form/securityComplianceCenter`
 
 ### Features
 
 - Users can now create FalconStor VTL instances from the page `/form/vtl`
 - Users can now set DNS Record Data to be the Primary IPV4 address for a VSI from the DNS Record form
 - Users can now create SCC V2 Resources fomr the `/form/sccV2` page
+- Users can now create additional VPC address prefixes from the VPN Gateways page
+- Users can now add VPN Gateway connections from the VPN Gateways page
 
 ### Fixes